Change license to MIT license
The Common Public License Version 1.0 was chosen for System Rules
because it was the same license like JUnit. Therefore users that are
already using JUnit don't need to accept a second kind of license.
Although it was very clear that with the same type of license System
Rules can depend on JUnit.

System Lambda is independent from JUnit. Therefore we can choose a
simpler license like the MIT license.