Extract code that checks that streams are restored
Reduce duplicate code and improve readability. I think that the test
classes don't need to store the details for checking that
System.err/System.out is restored. A nested class containing the tests
is sufficient. It is a level of indirection that I'm comfortable with.
The details of the tests are not important to me when I read the test
class.